moved tssPubKeyUtils to test/utils since it's only used for testing
